# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SplitTab – bill-splitting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List all groups
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ── Members ─────────────────────────────────────────────────────────────────
  /groups/{groupId}/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List members of a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: addMember
      tags: [members]
      summary: Add a member to a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /groups/{groupId}/members/{memberId}:
    delete:
      operationId: removeMember
      tags: [members]
      summary: Remove a member from a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
        - name: memberId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ── Expenses ─────────────────────────────────────────────────────────────────
...
  /groups/{groupId}/payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments in a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: recordPayment
      tags: [payments]
      summary: Record a settlement payment
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /groups/{groupId}/payments/{paymentId}:
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment record
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
        - name: paymentId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ── Balances & Settlements (wow) ──────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Group ──────────────────────────────────────────────────────────────────
    Group:
      type: object
      required: [id, name, currency, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        currency:
          type: string
        createdAt:
          type: string
    GroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        currency:
          type: string
          default: USD
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        currency:
          type: string

    # ── Member ─────────────────────────────────────────────────────────────────
    Member:
      type: object
      required: [id, groupId, name, createdAt]
      properties:
        id:
          type: number
        groupId:
          type: number
        name:
          type: string
        email:
          type: ["string", "null"]
        createdAt:
          type: string
    MemberInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string

    # ── Expense ────────────────────────────────────────────────────────────────
...
    Payment:
      type: object
      required: [id, groupId, fromMemberId, toMemberId, amount, date, createdAt]
      properties:
        id:
          type: number
        groupId:
          type: number
        fromMemberId:
          type: number
        fromMemberName:
          type: string
        toMemberId:
          type: number
        toMemberName:
          type: string
        amount:
          type: number
        date:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [fromMemberId, toMemberId, amount, date]
      properties:
        fromMemberId:
          type: number
        toMemberId:
          type: number
        amount:
          type: number
        date:
          type: string
        notes:
          type: string

    # ── Balances ────────────────────────────────────────────────────────────────
...