# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SplitTab – bill-splitting API
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: number;
  groupId: number;
  fromMemberId: number;
  fromMemberName?: string;
  toMemberId: number;
  toMemberName?: string;
  amount: number;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}